+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ### 2025-07-24 
   - inisiasi repo public
-  - perbaikan halamana profile
+  - perbaikan halaman profile
+  - perbaikan halaman user
+  - penambahan menu roles dan fitur crud role